Fielding Checks
Added the fielding checks.

Next will be adding stamina to skill level and decreasing stamina for players as effort is exerted, and recover stamina as time passes.

Adding bonus for pace bowlers in the beginning and end of the game.

Also adding a bowler selector, which chooses the best bowler for the over.

Then finally balancing the engine so it produces scores that would reflect a typical T20 Game